move component around to better locations
