+++ src/components/DeleteBtn.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import constants from 'utils/strings/constants';
-import { IconWithMessage } from './pages/gallery/SelectedFileOptions/GalleryOptions';
+import { IconWithMessage } from './IconWithMessage';
 
 const Wrapper = styled.button`
     border: none;
